newsletter/requirements.txt
google-generativeai>=0.8.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.120.0
python-dotenv>=1.0.0
Pillow>=10.0.0
